Fix other `buildarr compose` tests on Windows, remove `no cover` and `no branch` pragmas
